move jcfg to preferences module
